+++ app/src/desktopMain/kotlin/com/woowla/ghd/presentation/viewmodels/ReposToCheckViewModel.kt
@@ -11,6 +11,9 @@ import com.woowla.ghd.domain.services.RepoToCheckService
 import com.woowla.ghd.domain.entities.Event
 import com.woowla.ghd.core.eventbus.EventBus
 import com.woowla.ghd.core.utils.FlowReduxViewModel
+import com.woowla.ghd.domain.entities.RepoToCheckFilters
+import com.woowla.ghd.domain.entities.filtersRepoToCheck
+import com.woowla.ghd.domain.entities.groupNames
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 
 @OptIn(ExperimentalCoroutinesApi::class)
@@ -48,11 +51,14 @@ class ReposToCheckStateMachine(
                 on<Act.Reload> { _, state ->
                     load(state, searchQuery = state.snapshot.searchQuery)
                 }
-                on<Act.GroupNameFilterSelected> { action, state ->
-                    filter(state, groupName = action.groupName, isSelected = action.isSelected)
-                }
                 on<Act.SearchQueryChanged> { action, state ->
                     filter(state, searchQuery = action.searchQuery)
+                }
+                on<Act.GroupNameFilterChanged> { action, state ->
+                    filter(state, groupName = action.groupName, isSelected = action.isSelected)
+                }
+                on<Act.FiltersChanged> { action, state ->
+                    filter(state, filters = action.filters)
                 }
                 on<Act.DeleteRepoToCheck> { action, state ->
                     deleteRepo(state, repoToCheck = action.repoToCheck)
@@ -69,31 +75,29 @@ class ReposToCheckStateMachine(
     private suspend fun <T : St> load(state: State<T>, searchQuery: String): ChangedState<St> {
         return try {
             val appSettings = appSettingsService.get().getOrThrow()
-            val groupNameFiltersSelected = appSettings.filtersRepoToCheckGroupName
 
             repoToCheckService.getAll()
                 .fold(
                     onSuccess = { reposToCheck ->
-                        val groupNameFilters = reposToCheck.mapNotNull { it.groupName }.distinct().filter { it.isNotBlank() }.toSet()
-                        val groupNameFilterSizes = groupNameFilters.associateWith { groupName ->
+                        val allGroupNames = reposToCheck.mapNotNull { it.groupName }.distinct().filter { it.isNotBlank() }.toSet()
+                        val allGroupNamesSizes = allGroupNames.associateWith { groupName ->
                             reposToCheck.count { it.groupName == groupName }
                         }
-                        // clean up in case a group is no-available anymore
-                        val groupNameFiltersSelectedRecalculated = groupNameFiltersSelected.filter { groupNameFilters.contains(it) }.toSet()
 
-                        val releasesFiltered = reposToCheck.filter {
-                            groupNameFiltersSelectedRecalculated.isEmpty() || groupNameFiltersSelectedRecalculated.contains(it.groupName)
-                        }
+                        val releasesFiltered = reposToCheck.filter(
+                            searchQuery = searchQuery,
+                            filters = appSettings.filtersRepoToCheck,
+                        )
 
                         state.override {
                             St.Success(
                                 reposToCheck = reposToCheck,
                                 reposToCheckFiltered = releasesFiltered,
                                 appSettings = appSettings,
-                                groupNameFilters = groupNameFilters,
-                                groupNameFilterSizes = groupNameFilterSizes,
-                                groupNameFiltersSelected = groupNameFiltersSelectedRecalculated,
+                                allGroupNames = allGroupNames,
+                                allGroupNamesSizes = allGroupNamesSizes,
                                 searchQuery = searchQuery,
+                                filtersRepoToCheck = appSettings.filtersRepoToCheck,
                             )
                         }
                     },
@@ -114,31 +118,49 @@ class ReposToCheckStateMachine(
     private fun filter(state: State<St.Success>, searchQuery: String): ChangedState<St> {
         val reposToCheckFiltered = state.snapshot.reposToCheck.filter(
             searchQuery = searchQuery,
-            groupNames = state.snapshot.groupNameFiltersSelected,
+            filters = state.snapshot.filtersRepoToCheck,
         )
         return state.mutate {
             this.copy(
                 reposToCheckFiltered = reposToCheckFiltered,
                 searchQuery = searchQuery,
+                filtersRepoToCheck = filtersRepoToCheck,
             )
         }
     }
 
     private suspend fun filter(state: State<St.Success>, groupName: String, isSelected: Boolean): ChangedState<St> {
         val groupNameFiltersSelected = if (isSelected) {
-            state.snapshot.groupNameFiltersSelected - groupName
+            state.snapshot.filtersRepoToCheck.groupNames - groupName
         } else {
-            state.snapshot.groupNameFiltersSelected + groupName
+            state.snapshot.filtersRepoToCheck.groupNames + groupName
         }
 
-        val appSettings = state.snapshot.appSettings.copy(filtersRepoToCheckGroupName = groupNameFiltersSelected)
+        // clean up in case a group is no-available anymore
+        val allGroupNames = state.snapshot.allGroupNames
+        val groupNameFiltersSelectedRecalculated = groupNameFiltersSelected.filter { allGroupNames.contains(it) }.toSet()
+
+        val appSettings = AppSettings.filtersRepoToCheck.groupNames.modify(state.snapshot.appSettings) { groupNameFiltersSelectedRecalculated }
+
         appSettingsService.save(appSettings)
 
         // no change because the SETTINGS_UPDATED will be triggered automatically
         return state.noChange()
     }
 
-    private fun List<RepoToCheck>.filter(searchQuery: String, groupNames: Set<String>): List<RepoToCheck> {
+    private suspend fun filter(state: State<St.Success>, filters: RepoToCheckFilters): ChangedState<St> {
+        val appSettings = AppSettings.filtersRepoToCheck.modify(state.snapshot.appSettings) { filters }
+
+        appSettingsService.save(appSettings)
+
+        // no change because the SETTINGS_UPDATED will be triggered automatically
+        return state.noChange()
+    }
+
+    private fun List<RepoToCheck>.filter(
+        searchQuery: String,
+        filters: RepoToCheckFilters,
+    ): List<RepoToCheck> {
         val searchQueryIsBlank = searchQuery.isBlank()
         val searchQuernContainsName = { repoToCheck: RepoToCheck -> repoToCheck.repository?.name?.contains(searchQuery, ignoreCase = true) ?: false  }
         val searchQueryContainsGroup = { repoToCheck: RepoToCheck -> repoToCheck.groupName?.contains(searchQuery, ignoreCase = true) ?: false }
@@ -146,14 +168,40 @@ class ReposToCheckStateMachine(
             searchQueryIsBlank || searchQuernContainsName(repoToCheck) || searchQueryContainsGroup(repoToCheck)
         }
 
+        val groupNames = filters.groupNames
         val groupNamesAreEmpty = groupNames.isEmpty()
         val groupNamesContainsGroup = { repoToCheck: RepoToCheck -> groupNames.contains(repoToCheck.groupName) }
         val groupNameFilter = { repoToCheck: RepoToCheck ->
             groupNamesAreEmpty || groupNamesContainsGroup(repoToCheck)
         }
 
+        val otherFiltersFilter = otherFiltersFilter@{ repoToCheck: RepoToCheck ->
+            // no filter active
+            if (filters.anyFilterActive().not()) {
+                return@otherFiltersFilter true
+            }
+            // at least one filter active
+            var matches = false
+            if (filters.pullRequestSyncEnabled) {
+                matches = matches || repoToCheck.arePullRequestsEnabled
+            }
+            if (filters.pullRequestNotificationsEnabled) {
+                matches = matches || repoToCheck.arePullRequestsNotificationsEnabled
+            }
+            if (filters.pullRequestBranchFilterActive) {
+                matches = matches || !repoToCheck.pullBranchRegex.isNullOrBlank()
+            }
+            if (filters.releasesSyncEnabled) {
+                matches = matches || repoToCheck.areReleasesEnabled
+            }
+            if (filters.releasesNotificationsEnabled) {
+                matches = matches || repoToCheck.areReleasesNotificationsEnabled
+            }
+            matches
+        }
+
         return this.filter { repoToCheck ->
-            searchQueryFilter(repoToCheck) && groupNameFilter(repoToCheck)
+            searchQueryFilter(repoToCheck) && groupNameFilter(repoToCheck) && otherFiltersFilter(repoToCheck)
         }
     }
 
@@ -164,9 +212,9 @@ class ReposToCheckStateMachine(
             val reposToCheckFiltered: List<RepoToCheck>,
             val appSettings: AppSettings,
             val searchQuery: String,
-            val groupNameFilters: Set<String>,
-            val groupNameFilterSizes: Map<String, Int>,
-            val groupNameFiltersSelected: Set<String>,
+            val allGroupNames: Set<String>,
+            val allGroupNamesSizes: Map<String, Int>,
+            val filtersRepoToCheck: RepoToCheckFilters,
         ) : St
         data class Error(val throwable: Throwable) : St
     }
@@ -174,7 +222,8 @@ class ReposToCheckStateMachine(
     sealed interface Act {
         data object Reload : Act
         data class DeleteRepoToCheck(val repoToCheck: RepoToCheck) : Act
-        data class GroupNameFilterSelected(val isSelected: Boolean, val groupName: String) : Act
         data class SearchQueryChanged(val searchQuery: String) : Act
+        data class GroupNameFilterChanged(val isSelected: Boolean, val groupName: String) : Act
+        data class FiltersChanged(val filters: RepoToCheckFilters) : Act
     }
 }